Atualizado arquivos auxiliares

+++ ufla_201901/arq01/projeto_final/arquivos_auxiliares_201901/isummation.c
@@ -1,7 +1,7 @@
 
 #define SIZE 10
-unsigned short v[SIZE] = {5,2,9,6,3,10,7,4,1,8};
-unsigned short summ;
+int v[SIZE] = {5,2,9,6,3,10,7,4,1,8};
+int summ;
 
 void main (void) {
 	int i;
